Made mutations skip hediff base tick.
