+++ Source/Pawnmorphs/Esoteria/Hediffs/Hediff_StageChanges.cs
@@ -18,7 +18,8 @@ namespace Pawnmorph.Hediffs
         private int cachedStageIndex = -1;
         [Unsaved] private HediffStage cachedStage;
 
-        private List<Abilities.MutationAbility> abilities = new List<Abilities.MutationAbility>();
+        protected bool TickBase = true;
+
         private List<IStageChangeObserverComp> observerComps;
         private IEnumerable<IStageChangeObserverComp> ObserverComps
         {
@@ -66,7 +67,8 @@ namespace Pawnmorph.Hediffs
         /// </summary>
         public override void Tick()
         {
-            base.Tick();
+            if (TickBase)
+                base.Tick();
 
             int stageIndex = base.CurStageIndex; // Make sure to get the actual index from the base
             if (stageIndex != cachedStageIndex)
@@ -74,45 +76,10 @@ namespace Pawnmorph.Hediffs
                 var oldStage = cachedStage;
                 RecacheStage(stageIndex);
                 OnStageChanged(oldStage, cachedStage);
-                GenerateAbilities(cachedStage);
 
                 foreach (var comp in ObserverComps)
                     comp.OnStageChanged(oldStage, cachedStage);
             }
-
-            foreach (Abilities.MutationAbility ability in abilities)
-            {
-                ability.Tick();
-            }
-        }
-
-        private void GenerateAbilities(HediffStage stage)
-        {
-            if (stage is MutationStage mutationStage)
-            {
-                abilities = new List<Abilities.MutationAbility>();
-                if (mutationStage.abilities == null || mutationStage.abilities.Count == 0)
-                    return;
-
-                //Abilities.MutationAbility ability;
-                foreach (Abilities.MutationAbilityDef abilityDef in mutationStage.abilities)
-                {
-                    if (abilityDef.abilityClass.BaseType == typeof(Abilities.MutationAbility))
-                    {
-                        Abilities.MutationAbility ability = (Abilities.MutationAbility)Activator.CreateInstance(abilityDef.abilityClass, abilityDef);
-                        abilities.Add(ability);
-                        ability.Initialize(pawn);
-                    }
-                }
-            }
-        }
-
-        public override IEnumerable<Gizmo> GetGizmos()
-        {
-            foreach (Gizmo gizmo in base.GetGizmos()) yield return gizmo;
-
-            foreach (Abilities.MutationAbility item in abilities)
-                yield return item.Gizmo;
         }
 
         /// <summary>
@@ -137,17 +104,10 @@ namespace Pawnmorph.Hediffs
         {
             base.ExposeData();
             Scribe_Values.Look(ref cachedStageIndex, nameof(cachedStageIndex), base.CurStageIndex);
-            Scribe_Collections.Look(ref abilities, nameof(abilities));
 
             if (Scribe.mode == LoadSaveMode.PostLoadInit)
             {
                 RecacheStage(cachedStageIndex);
-
-                // Null if not previously saved.
-                if (abilities == null)
-                    abilities = new List<Abilities.MutationAbility>();
-
-                GenerateAbilities(cachedStage);
             }
         }
     }
